Moved Java related plugin to specific package
