+++ src/main/resources/META-INF/gradle-plugins/org.shipkit.java.properties
@@ -1,1 +1,1 @@
-implementation-class=org.shipkit.internal.gradle.ShipkitJavaPlugin
+implementation-class=org.shipkit.internal.gradle.java.ShipkitJavaPlugin
