+++ src/test/groovy/org/shipkit/internal/gradle/java/ShipkitJavaPluginTest.groovy
@@ -1,4 +1,4 @@
-package org.shipkit.internal.gradle
+package org.shipkit.internal.gradle.java
 
 import testutil.PluginSpecification
